# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive social media backend with posts, video calling, and
    file uploads
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Social Media Backend API
  version: "2.0"
paths:
...
  /profile:
    get:
      consumes:
      - application/json
      description: Get current user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Profile not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get my profile
      tags:
      - Users
    post:
      consumes:
      - multipart/form-data
      description: Create or update profile for authenticated user using multipart
        form data
      parameters:
      - description: First name
        in: formData
        name: first_name
        required: true
        type: string
      - description: Last name
        in: formData
        name: last_name
        required: true
        type: string
      - description: Display name
        in: formData
        name: display_name
        required: true
        type: string
      - description: Bio
        in: formData
        name: bio
        type: string
      - description: Phone number
        in: formData
        name: phone
        type: string
      - description: Date of birth (YYYY-MM-DD)
        in: formData
        name: date_of_birth
        type: string
      - description: Avatar image
        in: formData
        name: avatar
        type: file
      - description: Wall image
        in: formData
        name: wall_image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Created or updated profile
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create or update profile
      tags:
      - Users
    put:
      consumes:
      - multipart/form-data
      description: Update current user's profile information
      parameters:
      - description: First name
        in: formData
        name: first_name
        type: string
      - description: Last name
        in: formData
        name: last_name
        type: string
      - description: Display name
        in: formData
        name: display_name
        type: string
      - description: Bio
        in: formData
        name: bio
        type: string
      - description: Phone number
        in: formData
        name: phone
        type: string
      - description: Date of birth (YYYY-MM-DD)
        in: formData
        name: date_of_birth
        type: string
      - description: Avatar image
        in: formData
        name: avatar
        type: file
      - description: Wall image
        in: formData
        name: wall_image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Updated profile
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...